backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.2.0",
    "django-filter>=25.2",
    "django-phonenumber-field[phonenumbers]>=8.4.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "faker>=40.4.0",
    "gunicorn>=24.1.1",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "requests>=2.32.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "model-bakery>=1.23.2",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
